wip
In the implemented control unit,  the ir_write output signal is not high when it should be.
This is an experiment where ir_write is kept high in the instruction_decode stage (which kind of makes sense, really)
The tests still don't pass, but it looks much closer to correct.